+++ app/src/main/java/com/example/comusenias/presentation/screen/camera/OverlayVIew.kt
@@ -1,31 +1,50 @@
 import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Paint
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.comusenias.domain.models.ResultOverlayView
 import com.google.mediapipe.tasks.vision.core.RunningMode
 import com.google.mediapipe.tasks.vision.gesturerecognizer.GestureRecognizerResult
 import com.google.mediapipe.tasks.vision.handlandmarker.HandLandmarker
 import kotlin.math.max
 import kotlin.math.min
 
+
+
 @Composable
 fun OverlayView(
-    gestureRecognizerResults: List<GestureRecognizerResult>,
+    gestureRecognizerResults: ResultOverlayView,
     imageHeight: Int,
     imageWidth: Int,
     runningMode: RunningMode = RunningMode.IMAGE
 ) {
-    Canvas(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        for (result in gestureRecognizerResults) {
-            val gestures = result.gestures()
-            val firstGesture = gestures.getOrNull(0)
-            val category = firstGesture?.get(0)?.categoryName() ?: "none"
+    val result = gestureRecognizerResults.result
+    val width = gestureRecognizerResults.inputImageWidth
+    val height = gestureRecognizerResults.inputImageHeight
+
+    result.forEach {
+        val gestures = it.gestures()
+        val firstGesture = gestures.getOrNull(0)
+        val category = firstGesture?.get(0)?.categoryName() ?: "none"
+        val landmarksResult = it.landmarks()
+
+
+        PutTextCategory(text = category)
+
+        Canvas(
+            modifier = Modifier.fillMaxSize()
+        ) {
 
             val scaleFactor = when (runningMode) {
                 RunningMode.IMAGE, RunningMode.VIDEO -> {
@@ -50,7 +69,7 @@ fun OverlayView(
                 Color.Green // Si la categoría no es "None", establece el color en verde
             }
 
-            for (landmark in result.landmarks()) {
+            for (landmark in landmarksResult) {
                 for (normalizedLandmark in landmark) {
                     val correctedX = (normalizedLandmark.x() * imageWidth * scaleFactor)
                     val correctedY = (normalizedLandmark.y() * imageHeight * scaleFactor)
@@ -65,8 +84,8 @@ fun OverlayView(
                 var allConnectionsDetected = true
 
                 HandLandmarker.HAND_CONNECTIONS.forEach {
-                    val start = result.landmarks()[0].get(it!!.start())
-                    val end = result.landmarks()[0].get(it.end())
+                    val start = landmarksResult[0].get(it!!.start())
+                    val end = landmarksResult[0].get(it.end())
 
                     val startX = (start.x() * imageWidth * scaleFactor)
                     val startY = (start.y() * imageHeight * scaleFactor)
@@ -94,8 +113,35 @@ fun OverlayView(
                     )
                 }
             }
+
         }
     }
+
+
+}
+
+@Composable
+fun PutTextCategory(text: String) {
+    if (text == "none" || text == "") {
+        Text(
+            text = "Incorrecto", // Puedes personalizar este mensaje
+            fontSize = 20.sp,
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp),
+            color = Color.Red // Puedes personalizar el color
+        )
+    } else {
+        Text(
+            text = "Letra: $text", // Puedes personalizar este mensaje
+            fontSize = 20.sp,
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp),
+            color = Color.Green // Puedes personalizar el color
+        )
+    }
+
 }
 
 private const val LANDMARK_STROKE_WIDTH = 8f
